refactor: clean up http_mocker by removing unused functions and improving test case handling
