+++ internal/testutil/unit/http_mocker_data_test.go
@@ -34,12 +34,12 @@ func TestMockHTTPData_AddRoundtrip(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, "g1", mockData.Variables["groupId"])
 	require.Equal(t, "c1", mockData.Variables["clusterName"])
-	err = mockData.AddRoundtrip(t, rt, false)
-	require.Error(t, err)
-	changeError, ok := err.(*unit.VariablesChangedError)
-	require.True(t, ok)
-	require.Len(t, changeError.Changes, 2)
-	assert.Equal(t, map[string]string{"groupId": "groupId2", "clusterName": "clusterName2"}, changeError.ChangedNamesMap())
-	assert.Equal(t, map[string]string{"g1": "g1", "c1": "c1"}, changeError.ChangedValuesMap())
-	assert.Equal(t, map[string]string{"clusterName": "c1", "clusterName2": "c1", "groupId": "g1", "groupId2": "g1"}, mockData.Variables)
+	rt2 := &unit.RoundTrip{
+		Variables:  map[string]string{"groupId": "g2", "clusterName": "c2"},
+		StepNumber: 1,
+		Request:    unit.RequestInfo{},
+	}
+	err = mockData.AddRoundtrip(t, rt2, false)
+	require.NoError(t, err)
+	assert.Equal(t, map[string]string{"clusterName": "c1", "clusterName2": "c2", "groupId": "g1", "groupId2": "g2"}, mockData.Variables)
 }
